requirements.txt
streamlit>=1.31
networkx>=3.2
python-chess>=1.999
